Update from 1.11.x

+++ public/plugin/bbb/lib/bbb_plugin.class.php
@@ -7,12 +7,10 @@
  * main/img/icons/64/plugin_name.png
  * main/img/icons/64/plugin_name_na.png
 */
-/**
- * Videoconference plugin with BBB.
- */
 
 /**
- * Class BBBPlugin.
+ * Class BBBPlugin
+ * Videoconference plugin with BBB
  */
 class BBBPlugin extends Plugin
 {
@@ -22,6 +20,10 @@ class BBBPlugin extends Plugin
     const LAUNCH_TYPE_DEFAULT = 0;
     const LAUNCH_TYPE_SET_BY_TEACHER = 1;
     const LAUNCH_TYPE_SET_BY_STUDENT = 2;
+
+    const ROOM_OPEN = 0;
+    const ROOM_CLOSE = 1;
+    const ROOM_CHECK = 2;
 
     public $isCoursePlugin = true;
 
@@ -47,8 +49,8 @@ class BBBPlugin extends Plugin
     protected function __construct()
     {
         parent::__construct(
-            '2.8.1',
-            'Julio Montoya, Yannick Warnier, Angel Fernando Quiroz Campos',
+            '2.8.2',
+            'Julio Montoya, Yannick Warnier, Angel Fernando Quiroz Campos, Jose Angel Ruiz',
             [
                 'tool_enable' => 'boolean',
                 'host' => 'text',
@@ -63,51 +65,38 @@ class BBBPlugin extends Plugin
                     'type' => 'select',
                     'options' => [
                         PLATFORM_ADMIN => get_lang('Administrator'),
-                        COURSEMANAGER => get_lang('Trainer'),
-                        STUDENT => get_lang('Learner'),
-                        STUDENT_BOSS => get_lang('LearnerBoss'),
+                        COURSEMANAGER => get_lang('Teacher'),
+                        STUDENT => get_lang('Student'),
+                        STUDENT_BOSS => get_lang('StudentBoss'),
                     ],
                     'attributes' => ['multiple' => 'multiple'],
                 ],
                 'interface' => [
                     'type' => 'select',
                     'options' => [
-                        self::INTERFACE_FLASH => 'Flash',
                         self::INTERFACE_HTML5 => 'HTML5',
+                        self::INTERFACE_FLASH => 'Flash',
                     ],
                 ],
                 'launch_type' => [
                     'type' => 'select',
                     'options' => [
                         self::LAUNCH_TYPE_DEFAULT => 'SetByDefault',
-                        self::LAUNCH_TYPE_SET_BY_TEACHER => 'SetByTrainer',
-                        self::LAUNCH_TYPE_SET_BY_STUDENT => 'SetByLearner',
+                        self::LAUNCH_TYPE_SET_BY_TEACHER => 'SetByTeacher',
+                        self::LAUNCH_TYPE_SET_BY_STUDENT => 'SetByStudent',
                     ],
                     'translate_options' => true, // variables will be translated using the plugin->get_lang
                 ],
                 'allow_regenerate_recording' => 'boolean',
+                // Default course settings, must be the same as $course_settings
+                'big_blue_button_record_and_store' => 'checkbox',
+                'bbb_enable_conference_in_groups' => 'checkbox',
+                'bbb_force_record_generation' => 'checkbox',
+                'disable_course_settings' => 'boolean',
             ]
         );
 
         $this->isAdminPlugin = true;
-    }
-
-    /**
-     * @param string $variable
-     *
-     * @return bool
-     */
-    public function validateCourseSetting($variable)
-    {
-        if ('bbb_enable_conference_in_groups' === $variable) {
-            if ('true' === $this->get('enable_conference_in_course_groups')) {
-                return true;
-            }
-
-            return false;
-        }
-
-        return true;
     }
 
     /**
@@ -117,11 +106,71 @@ class BBBPlugin extends Plugin
     {
         static $result = null;
 
-        return $result ?: $result = new self();
+        return $result ? $result : $result = new self();
     }
 
     /**
-     * Install.
+     * @param string $variable
+     *
+     * @return bool
+     */
+    public function validateCourseSetting($variable)
+    {
+        if ($this->get('disable_course_settings') === 'true') {
+            return false;
+        }
+
+        $result = true;
+        switch ($variable) {
+            case 'bbb_enable_conference_in_groups':
+                $result = $this->get('enable_conference_in_course_groups') === 'true';
+                break;
+            case 'bbb_force_record_generation':
+                $result = $this->get('allow_regenerate_recording') === 'true';
+                break;
+            case 'big_blue_button_record_and_store':
+        }
+
+        return $result;
+    }
+
+    /**
+     *
+     * @return array
+     */
+    public function getCourseSettings()
+    {
+        $settings = [];
+        if ($this->get('disable_course_settings') !== 'true') {
+            $settings = parent::getCourseSettings();
+        }
+
+        return $settings;
+    }
+
+    /**
+     *
+     * @return \Plugin
+     */
+    public function performActionsAfterConfigure()
+    {
+        $result = $this->get('disable_course_settings') === 'true';
+        if ($result) {
+            $valueConference = $this->get('bbb_enable_conference_in_groups') === 'true' ? 1 : 0;
+            self::update_course_field_in_all_courses('bbb_enable_conference_in_groups', $valueConference);
+
+            $valueForceRecordGeneration = $this->get('bbb_force_record_generation') === 'true' ? 1 : 0;
+            self::update_course_field_in_all_courses('bbb_force_record_generation', $valueForceRecordGeneration);
+
+            $valueForceRecordStore = $this->get('big_blue_button_record_and_store') === 'true' ? 1 : 0;
+            self::update_course_field_in_all_courses('big_blue_button_record_and_store', $valueForceRecordStore);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Install
      */
     public function install()
     {
@@ -147,19 +196,20 @@ class BBBPlugin extends Plugin
                 access_url INT NOT NULL DEFAULT 1,
                 video_url TEXT NULL,
                 has_video_m4v TINYINT NOT NULL DEFAULT 0,
-                interface INT NOT NULL DEFAULT 0                                              
+                interface INT NOT NULL DEFAULT 0
                 )";
         Database::query($sql);
 
         Database::query(
-            'CREATE TABLE IF NOT EXISTS plugin_bbb_room (
+            "CREATE TABLE IF NOT EXISTS plugin_bbb_room (
                 id int NOT NULL AUTO_INCREMENT PRIMARY KEY,
                 meeting_id int NOT NULL,
                 participant_id int(11) NOT NULL,
                 in_at datetime,
                 out_at datetime,
-                interface int NOT NULL DEFAULT 0
-            );'
+                interface int NOT NULL DEFAULT 0,
+                close INT NOT NULL DEFAULT 0
+            );"
         );
         $fieldLabel = 'plugin_bbb_course_users_limit';
         $fieldType = ExtraField::FIELD_TYPE_INTEGER;
@@ -206,7 +256,7 @@ class BBBPlugin extends Plugin
     }
 
     /**
-     * Uninstall.
+     * Uninstall
      */
     public function uninstall()
     {
@@ -232,41 +282,79 @@ class BBBPlugin extends Plugin
             'launch_type',
         ];
 
+        $urlId = api_get_current_access_url_id();
+
         foreach ($variables as $variable) {
-            $sql = "DELETE FROM $t_settings WHERE variable = '$variable'";
+            $sql = "DELETE FROM $t_settings WHERE variable = '$variable' AND access_url = $urlId";
             Database::query($sql);
         }
-        $extraField = new ExtraField('course');
-        $extraFieldInfo = $extraField->get_handler_field_info_by_field_variable(
-            'plugin_bbb_course_users_limit'
-        );
-        if (!empty($extraFieldInfo)) {
-            $extraField->delete($extraFieldInfo['id']);
-        }
-        $extraField = new ExtraField('session');
-        $extraFieldInfo = $extraField->get_handler_field_info_by_field_variable(
-            'plugin_bbb_session_users_limit'
-        );
-        if (!empty($extraFieldInfo)) {
-            $extraField->delete($extraFieldInfo['id']);
+
+        $em = Database::getManager();
+        $sm = $em->getConnection()->getSchemaManager();
+        if ($sm->tablesExist('plugin_bbb_meeting')) {
+            Database::query("DELETE FROM plugin_bbb_meeting WHERE access_url = $urlId");
         }
 
-        $sql = "DELETE FROM $t_options WHERE variable = 'bbb_plugin'";
-        Database::query($sql);
+        // Only delete tables if it's uninstalled from main url.
+        if (1 == $urlId) {
+            $extraField = new ExtraField('course');
+            $extraFieldInfo = $extraField->get_handler_field_info_by_field_variable(
+                'plugin_bbb_course_users_limit'
+            );
+            if (!empty($extraFieldInfo)) {
+                $extraField->delete($extraFieldInfo['id']);
+            }
+            $extraField = new ExtraField('session');
+            $extraFieldInfo = $extraField->get_handler_field_info_by_field_variable(
+                'plugin_bbb_session_users_limit'
+            );
+            if (!empty($extraFieldInfo)) {
+                $extraField->delete($extraFieldInfo['id']);
+            }
 
-        // hack to get rid of Database::query warning (please add c_id...)
-        $sql = "DELETE FROM $t_tool WHERE name = 'bbb' AND c_id != 0";
-        Database::query($sql);
+            $sql = "DELETE FROM $t_options WHERE variable = 'bbb_plugin'";
+            Database::query($sql);
 
-        Database::query('DROP TABLE IF EXISTS plugin_bbb_room');
-        Database::query('DROP TABLE IF EXISTS plugin_bbb_meeting');
+            // hack to get rid of Database::query warning (please add c_id...)
+            $sql = "DELETE FROM $t_tool WHERE name = 'bbb' AND c_id != 0";
+            Database::query($sql);
 
-        // Deleting course settings
-        $this->uninstall_course_fields_in_all_courses($this->course_settings);
+            if ($sm->tablesExist('plugin_bbb_room')) {
+                Database::query('DROP TABLE IF EXISTS plugin_bbb_room');
+            }
+            if ($sm->tablesExist('plugin_bbb_meeting')) {
+                Database::query('DROP TABLE IF EXISTS plugin_bbb_meeting');
+            }
+
+            // Deleting course settings
+            $this->uninstall_course_fields_in_all_courses($this->course_settings);
+        }
     }
 
     /**
-     * Return an array with URL.
+     * Update
+     */
+    public function update()
+    {
+        $sql = "SHOW COLUMNS FROM plugin_bbb_room WHERE Field = 'close'";
+        $res = Database::query($sql);
+
+        if (Database::num_rows($res) === 0) {
+            $sql = "ALTER TABLE plugin_bbb_room ADD close int unsigned NULL";
+            $res = Database::query($sql);
+            if (!$res) {
+                echo Display::return_message($this->get_lang('ErrorUpdateFieldDB'), 'warning');
+            }
+
+            Database::update(
+                'plugin_bbb_room',
+                ['close' => BBBPlugin::ROOM_CLOSE]
+            );
+        }
+    }
+
+    /**
+     * Return an array with URL
      *
      * @param string $conferenceUrl
      *
@@ -287,11 +375,13 @@ class BBBPlugin extends Plugin
      */
     public function getFlashUrl($conferenceUrl)
     {
-        return [
+        $data = [
             'text' => $this->get_lang('EnterConferenceFlash'),
             'url' => $conferenceUrl.'&interface='.self::INTERFACE_FLASH,
             'icon' => 'resources/img/64/videoconference_flash.png',
         ];
+
+        return $data;
     }
 
     /**
@@ -301,10 +391,34 @@ class BBBPlugin extends Plugin
      */
     public function getHtmlUrl($conferenceUrl)
     {
-        return [
+        $data = [
             'text' => $this->get_lang('EnterConferenceHTML5'),
             'url' => $conferenceUrl.'&interface='.self::INTERFACE_HTML5,
             'icon' => 'resources/img/64/videoconference_html5.png',
         ];
+
+        return $data;
+    }
+
+    /**
+     * Set the course setting in all courses
+     *
+     * @param bool $variable Course setting to update
+     * @param bool $value New values of the course setting
+     */
+    public function update_course_field_in_all_courses($variable, $value)
+    {
+        // Update existing courses to add the new course setting value
+        $table = Database::get_main_table(TABLE_MAIN_COURSE);
+        $sql = "SELECT id FROM $table ORDER BY id";
+        $res = Database::query($sql);
+        $courseSettingTable = Database::get_course_table(TABLE_COURSE_SETTING);
+        while ($row = Database::fetch_assoc($res)) {
+            Database::update(
+                $courseSettingTable,
+                ['value' => $value],
+                ['variable = ? AND c_id = ?' => [$variable, $row['id']]]
+            );
+        }
     }
 }
